Give the millisecond clock and nVars() fixed-width types
"long" is 64 bits on LP64 but 32 on ILP32 and LLP64, so these two
interfaces meant different things depending on the target.

getCurrentTime() returned a millisecond count in a long. Where time_t
is 32 bits -- the default on i386 -- "1000 * t.tv_sec" overflowed
before the return type was reached, which is undefined, and the result
did not fit a 32-bit long anyway. It survived only because every caller
takes differences, and the wraparound cancels for short intervals.

Widen tv_sec before the multiply and return int64_t. RunTimes' Element,
times map, addTime() and lastTime follow, as do print()'s millisecond
accumulators, which were int while summing 64-bit values.

The function had two byte-identical definitions, in RunTimes.cpp and
ASTmisc.cpp. Keep the one in Util, where the other timing code lives,
and move the declaration to RunTimes.h. No include changes are needed:
AST.h already reaches RunTimes.h through UsefulDefs.h.

nVars() returned unsigned long from every solver. All five underlying
solvers return int or uint32_t, and the one caller assigns straight
back to an int, so the width was 64 bits on Linux, 32 on i386 and
Win64, and meaningful nowhere. Make it uint32_t.

rewrite_rule_gen's RewriteRule::setVerified() takes an int, so the
elapsed durations passed to it get an explicit cast rather than an
implicit narrowing.

+++ include/stp/Sat/SATSolver.h
@@ -172,7 +172,7 @@ public:
 
   virtual uint32_t newVar() = 0;
 
-  virtual unsigned long nVars() const = 0;
+  virtual uint32_t nVars() const = 0;
 
   virtual void printStats() const = 0;
